pyproject.toml
[project]
name = "cmp"
version = "0.1.0"
description = "CLI Music Player - A terminal-based music player with visualizations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "CLI Music Player"}
]
dependencies = [
    "textual>=0.41.0",
    "pyaudio>=0.2.11",
    "soundfile>=0.12.0",
    "numpy>=1.21.0",
    "mutagen>=1.47.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
music = "cmp.__main__:main"
cmp-music = "cmp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
